+++ lib/fluent_ui.dart
@@ -57,8 +57,8 @@ export 'src/controls/form/pickers/time_picker.dart';
 export 'src/styles/motion/page_transitions.dart';
 export 'src/styles/acrylic.dart';
 export 'src/styles/color.dart';
-// export 'src/styles/icons.dart';
 export 'src/styles/theme.dart';
 export 'src/styles/typography.dart';
 
+export 'src/styles/focus.dart';
 export 'src/utils/header.dart';